fix: rename `commons::symbols` to `commons::scanner`
