+++ commons/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod config;
 pub mod log_id;
-pub mod symbols;
+pub mod scanner;
 
 #[cfg(feature = "test_runner")]
 pub mod test_runner;
