+++ core/src/elements/enclosed/verbatim.rs
@@ -7,7 +7,7 @@ use unimarkup_render::render::Render;
 use crate::elements::blocks::Block;
 use crate::elements::Blocks;
 use crate::parser::{ElementParser, TokenizeOutput};
-use unimarkup_commons::symbols::{Symbol, SymbolKind};
+use unimarkup_commons::scanner::{Symbol, SymbolKind};
 
 /// Structure of a Unimarkup verbatim block element.
 #[derive(Debug, PartialEq, Eq, Clone)]
